Add dynamic splat loader
