+++ packages/dev/loaders/src/dynamic.ts
@@ -1,7 +1,9 @@
 import { registerOBJLoader } from "./OBJ/objDynamicFileLoader";
 import { registerGLTFLoader } from "./glTF/glTFDynamicFileLoader";
+import { registerSPLATLoader } from "./SPLAT/splatDynamicFileLoader";
 
 export function registerLoaders() {
     registerOBJLoader();
     registerGLTFLoader();
+    registerSPLATLoader();
 }
